refactor, format, lint
